+++ CHANGES.md
@@ -28,6 +28,8 @@ To be released.
 
      -  An `Address` became represented as a single hexadecimal string in JSON.
      -  A `BlockHash` became represented as a single hexadecimal string in JSON.
+     -  A `HashDigest` became represented as a single hexadecimal string in
+        JSON.
 
 ### Bug fixes